requirements.txt
flask
tensorflow-cpu
tf-keras
numpy
opencv-python-headless
imutils
pillow